+++ libs/vds_p2p_network/private/udp_transport_session_p.h
@@ -91,7 +91,7 @@ namespace vds {
 
       auto curent = this->min_incoming_sequence_;
       uint32_t result = 0;
-      for(int i = 0; i < 8 * sizeof(result); ++i){
+      for(size_t i = 0; i < 8 * sizeof(result); ++i){
         ++curent;
         result >>= 1;
         if(this->future_data_.end() != this->future_data_.find(curent)){
